+++ activerecord/lib/active_record/database_configurations/database_config.rb
@@ -65,6 +65,10 @@ module ActiveRecord
       def for_current_env?
         env_name == ActiveRecord::ConnectionHandling::DEFAULT_ENV.call
       end
+
+      def schema_cache_path
+        raise NotImplementedError
+      end
     end
   end
 end