+++ activerecord/lib/active_record/database_configurations/hash_config.rb
@@ -87,6 +87,13 @@ module ActiveRecord
       def adapter
         configuration_hash[:adapter]
       end
+
+      # The path to the schema cache dump file for a database.
+      # If omitted, the filename will be read from ENV or a
+      # default will be derived.
+      def schema_cache_path
+        configuration_hash[:schema_cache_path]
+      end
     end
   end
 end